bug: toolbar button
We are registering the event listener on the toolbar button dynamically
from JS using addEventListener().  While the button is in the toolbar
pallette, it does not appear to be available on the chrome document
object until it has been installed.  The order of operations has been
changed to meet that scheme, but that means if a user removes the button
after it has been installed, the event will not be registered until
Firefox is started up again.

 On branch rc_0.1
	modified:   firefox/__init__0.1.js